server/requirements.txt
flask
sqlalchemy
python-dotenv